--- a/debian/.dockerignore
+++ b/debian/.dockerignore
@@ -1,0 +1,2 @@
+**/mmc_*.img
+**/mmc_*.img.xz