--- a/debian/nanopi-r5s/make_debian_img.sh
+++ b/debian/nanopi-r5s/make_debian_img.sh
@@ -1,5 +1,6 @@
 #!/bin/sh
 
-# For return codes, see common/make_debian_img_for_model.sh
-cd "$(dirname "$(realpath "$0")")"
-exec sh ../common/make_debian_img_for_model.sh r5s "$@"
+# For configuration, see ../env.sh
+# For return codes, see ../main.sh
+cd "$(dirname "$(realpath "$0")")/.."
+exec sh ./main.sh r5s "$@"