--- a/debian/steps/init_builder_A.sh
+++ b/debian/steps/init_builder_A.sh
@@ -1,0 +1,13 @@
+#!/bin/sh
+
+# Copyright (C) 2023, John Clark <inindev@gmail.com>
+# Copyright (C) 2024, Etienne Dublé <etienne.duble@imag.fr>
+
+set -e
+. ./env.sh
+. ./tools.sh
+
+print_hdr "preparing build environment"
+
+apt update
+apt install -y wget xz-utils debootstrap