--- a/debian/tools.sh
+++ b/debian/tools.sh
@@ -1,0 +1,24 @@
+#!/bin/sh
+
+# Copyright (C) 2023, John Clark <inindev@gmail.com>
+# Copyright (C) 2024, Etienne Dublé <etienne.duble@imag.fr>
+
+. ./env.sh
+
+is_param() {
+    local item match
+    for item in "$@"; do
+        if [ -z "$match" ]; then
+            match="$item"
+        elif [ "$match" = "$item" ]; then
+            return 0
+        fi
+    done
+    return 1
+}
+
+print_hdr() {
+    h1="${blu}==>${rst} ${bld}"
+    msg="$1"
+    echo "\n${h1}$msg...${rst}"
+}