… debian/common/make_debian_img_for_model.sh