@inindev 2024-06-17
Thank you for the outstanding contribution; overall, I like it. I like not requiring root, and it is nice that it can be run cross-platform. I had been looking at github actions as a next step and I want to better understand how this could all come together looking forward to that goal.  I was not able to look at this last week, give me a couple of days to better understand it.