@eduble 2024-06-17
> I had been looking at github actions as a next step and I want to better understand how this could all come together looking forward to that goal.

OK.
Something related: in the past I created the following docker image of raspberry pi OS and automated the build:
https://github.com/eduble/rpi-mini
Whenever I pushed to this github repo the image was automatically rebuilt and pushed to the docker hub here:
https://hub.docker.com/r/eduble/rpi-mini
However, if I recall well, in this case the automated build was processed by the docker hub cloud, not by github.
And I think this now requires a Pro account on the docker hub.

However, you may still find interesting information in the Dockerfile (https://github.com/eduble/rpi-mini/blob/master/Dockerfile), for automating the build of your image using github actions. Checkout the comment on line 20. This kind of tricks might be needed to run the image on the github cloud too.

For this problem of qemu emulation in the cloud context I mostly reused the work done by the guys at balena.io:
https://blog.balena.io/building-arm-containers-on-any-x86-machine-even-dockerhub/
(very useful blog post)

> I was not able to look at this last week, give me a couple of days to better understand it.

OK. You can take your time, this is not an urgent matter on my side.